Moved IsPathTemplate to resolve check-agent-deps failures. Added additional path template validation.
