+++ pilot/pkg/security/authz/matcher/template_test.go
@@ -20,8 +20,6 @@ import (
 	uri_template "github.com/envoyproxy/go-control-plane/envoy/extensions/path/match/uri_template/v3"
 	"github.com/google/go-cmp/cmp"
 	"google.golang.org/protobuf/testing/protocmp"
-
-	"istio.io/istio/pkg/test/util/assert"
 )
 
 func TestPathTemplateMatcher(t *testing.T) {
@@ -59,47 +57,6 @@ func TestPathTemplateMatcher(t *testing.T) {
 			if !cmp.Equal(got, tc.want, protocmp.Transform()) {
 				t.Errorf("want %v but got %v", tc.want, got)
 			}
-		})
-	}
-}
-
-func TestIsPathTemplate(t *testing.T) {
-	testCases := []struct {
-		name           string
-		path           string
-		isPathTemplate bool
-	}{
-		{
-			name:           "matchOneOnly",
-			path:           "foo/bar/{*}",
-			isPathTemplate: true,
-		},
-		{
-			name:           "matchOneOnly",
-			path:           "foo/{**}/bar",
-			isPathTemplate: true,
-		},
-		{
-			name:           "matchAnyAndOne",
-			path:           "{*}/bar/{**}",
-			isPathTemplate: true,
-		},
-		{
-			name:           "stringMatch",
-			path:           "foo/bar/*",
-			isPathTemplate: false,
-		},
-		{
-			name:           "namedVariable",
-			path:           "foo/bar/{buzz}",
-			isPathTemplate: false,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			pathTemplate := IsPathTemplate(tc.path)
-			assert.Equal(t, tc.isPathTemplate, pathTemplate)
 		})
 	}
 }
