+++ tests/integration/security/testdata/authz/path-templating.yaml.tmpl
@@ -10,4 +10,4 @@ spec:
   rules:
     - to:
         - operation:
-            paths: ["/allow/admin/{**}", "/foo/{*}/bar/{**}.txt"]
+            paths: ["/allow/admin/{**}", "/foo/{*}/bar/{**}.txt", "/store/{**}/cart"]
